remove veranjakanda
I discovered to my horror that the Veranjakanda was unaccountably repeated in the text, being both in pi-tv-bu-vb-pj1 and pi-tv-bu-vb-intro. I thought I had fixed this long ago. In any case, i have removed it from the beginning of Pj 1, and kept it as intro. This created the problem of the sc numbers in pj1 starting from 41. I have fixed this so they start from 1. However I have not fixed any cross references and data, so these must be adjusted accordingly. Sorry! Note that the sc numbers here used to be identical with the <a class="ms" id="p_1V_41">, etc numbers.