delete repeated verse and translate untranslated
In the original version, the first verse of this translation was repeated by mistake, and another verse (204) was omitted. The total verse count is thus unaffected, which is how it slipped through. I have eliminated the extra verse and supplied a new translation. I have also submitted the same correction to the publisher, who is printing this now. Hopefully it will be in time!